requirements/requirements.txt
matplotlib==3.2.1
numpy==1.18.0
torch==1.8.1
pickle==
pillow==8.1.0
tqdm==4.60.0
torchvision==0.9.1
sklearn==0.0